+++ pkg/front_end/test/macros/application/macro_application_test.dart
@@ -4,7 +4,7 @@
 
 import 'dart:io' show Directory, File, Platform;
 
-import 'package:_fe_analyzer_shared/src/macros/api.dart';
+import 'package:_fe_analyzer_shared/src/macros/api.dart' hide Library;
 import 'package:_fe_analyzer_shared/src/macros/executor.dart';
 import 'package:_fe_analyzer_shared/src/testing/id.dart'
     show ActualData, ClassId, Id, LibraryId;